+++ shared/events/edgeware/filters/enricher.ts
@@ -1,0 +1,7 @@
+import { ApiPromise } from '@polkadot/api';
+import { Event } from '@polkadot/types/interfaces';
+import { SubstrateEventType } from '../types';
+
+export function enrichEvent(api: ApiPromise, type: SubstrateEventType, event: Event): any {
+
+}
